Time: 11 ms (36.22%), Space: 23 MB (46.85%) - LeetHub

+++ 0875-koko-eating-bananas/0875-koko-eating-bananas.cpp
@@ -2,7 +2,7 @@ class Solution {
 public:
 
     bool is_pos(vector<int>& piles, int h, int k) {
-        int hrs = 0;
+        long long hrs = 0;
         for(int i = 0; i < piles.size(); i++) {
             hrs += (piles[i] + k - 1) / k;
         }
